fix(rules): port no-multi-comp faithfully
