+++ .changeset/strict-react-no-multi-comp.md
@@ -1,0 +1,6 @@
+---
+"oxlint-plugin-react-doctor": patch
+"react-doctor": patch
+---
+
+Make `react-doctor/no-multi-comp` faithfully enforce Oxlint's one-component-per-file contract, and preserve React Doctor's lower-noise default behavior under `react-doctor/no-crowded-component-file`.
