+++ packages/oxlint-plugin-react-doctor/src/plugin/liveness/liveness-fixtures.ts
@@ -728,6 +728,9 @@ export const livenessFixtures: Readonly<Record<string, LivenessFixture>> = {
   "no-cramped-container-padding": {
     code: 'const Panel = () => <div className="border rounded p-1">Status</div>;',
   },
+  "no-crowded-component-file": {
+    code: "const Foo = () => <div />; const Bar = () => <div />; const Baz = () => <div />;",
+  },
   "no-crushed-letter-spacing": {
     code: 'const Heading = () => <h1 style={{ letterSpacing: "-0.12em" }}>Readable heading</h1>;',
   },
